Update to newest mod versions + School armor now gives some general proficiency

+++ src/main/java/de/totodev/spellbooktweaks/ClientReserveManaData.java
@@ -16,7 +16,7 @@ public class ClientReserveManaData {
     }
 
     public static int getReserveMana() {
-        return playerReserveManaData.spellbookTweaks$getReserveMana();
+        return (int) playerReserveManaData.spellbookTweaks$getReserveMana();
     }
 
     public static void setReserveMana(float mana) {
